app-editor/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
python-multipart==0.0.6
google-generativeai==0.8.0
pysubs2==1.7.2
aiofiles==23.2.1
httpx==0.26.0
python-dotenv==1.0.0
pytest==8.0.0
